Move classes in subpackages
