+++ gradle.properties
@@ -2,7 +2,7 @@
 # -> https://www.jetbrains.org/intellij/sdk/docs/reference_guide/intellij_artifacts.html
 
 pluginGroup = com.github.pirocraft.challengingtimer
-pluginName_ = Challenging timer
+pluginName_ = Challenging Timer
 pluginVersion = 0.1.2
 pluginSinceBuild = 193
 pluginUntilBuild = 202.*
