+++ src/main/kotlin/com/github/pirocraft/challengingtimer/ui/TimerToolWindowFactory.kt
@@ -1,4 +1,4 @@
-package com.github.pirocraft.challengingtimer
+package com.github.pirocraft.challengingtimer.ui
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindow
